// shared/src/types/resources.d.ts
export type ResourceType = 'timber' | 'clay' | 'iron' | 'grain' | 'wool';
export type TerrainType = ResourceType | 'desert';
export type ResourceBundle = Record<ResourceType, number>;
export declare const EMPTY_RESOURCES: ResourceBundle;
export declare const ALL_RESOURCES: ResourceType[];
//# sourceMappingURL=resources.d.ts.map